style: Welcome 컴포넌트 스타일 분리
